wiring: move wiring into own folder
