+++ nightmare.vcxproj
@@ -253,7 +253,7 @@
     <ClCompile Include="src\tools\fire_projectile.cc" />
     <ClCompile Include="src\tools\remove_block.cc" />
     <ClCompile Include="src\tools\remove_surface.cc" />
-    <ClCompile Include="src\wiring.cc" />
+    <ClCompile Include="src\wiring\wiring.cc" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\blob.h" />
@@ -295,7 +295,7 @@
     <ClInclude Include="src\timer.h" />
     <ClInclude Include="src\tools.h" />
     <ClInclude Include="src\winunistd.h" />
-    <ClInclude Include="src\wiring.h" />
+    <ClInclude Include="src\wiring\wiring.h" />
     <ClInclude Include="winerr.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
