+++ src/wiring/wiring.h
@@ -2,7 +2,7 @@
 
 #include <glm/glm.hpp>
 
-#include "render_data.h"
+#include "../render_data.h"
 
 struct ship_space;
